Treat both the default graph pixel height and default coord height as integers. One of these is actually used as a double internally, but we should hide that detail from users since they have no need to know it.
git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@56 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ igb/src/com/affymetrix/igb/util/GraphGlyphUtils.java
@@ -61,7 +61,13 @@ public class GraphGlyphUtils {
   public static final String default_attach_mode = USE_CURRENT_HEIGHT;  // default mode if not specified in prefs
 
   public static final int default_pix_height = 60;
-  public static final double default_coord_height = 100;
+  
+  /** Default value for height of attached (non-floating) graphs.  Although
+   *  the height will ultimately have to be expressed as a double rather than
+   *  an integer, there is no good reason to bother the users with that detail,
+   *  so the default should be treated as an integer.
+   */
+  public static final int default_coord_height = 100;
   public static final boolean default_use_floating_graphs = false;
 
   public static final Color[] default_graph_colors = 
@@ -115,7 +121,7 @@ public class GraphGlyphUtils {
       //    but setting height will matter
       double yheight = coordbox.height;
       if (use_fixed_coord_height) {
-        yheight = getGraphPrefsNode().getDouble(PREF_ATTACHED_COORD_HEIGHT, default_coord_height);
+        yheight = getGraphPrefsNode().getDouble(PREF_ATTACHED_COORD_HEIGHT, (double) default_coord_height);
       }
 
       gl.setCoords(tempbox.x, coordbox.y, tempbox.width, yheight);
